Fixes #485 - Add `--single-version-externally-managed` to setup.py command line, 'cause it seems to help (#484)
There is some magic. When installing to system `site_packages`, things just work. When installing to these other custom locations, things do not work.

My installs were broken even after I reverted https://github.com/skupperproject/skupper-router/issues/440. Either I was reverting wrong, or this thing is just broken, regardless whether setuptools or distutils is imported in setup.py

I was reading about the .pth files, and all the .egg-info stuff and so on. Did not help me to make sense out of it. https://setuptools.pypa.io/en/latest/deprecated/easy_install.html#custom-installation-locations

> /home/jdanek/.cache/pypoetry/virtualenvs/skupper-router-7_an9NB3-py3.10/lib/python3.10/site-packages/pkg_resources/__init__.py:122: PkgResourcesDeprecationWarning: 2.0.1-32-g535900fe is an invalid version and will not be supported in a future release

> /home/jdanek/.cache/pypoetry/virtualenvs/skupper-router-7_an9NB3-py3.10/lib/python3.10/site-packages/setuptools/command/install.py:34: SetuptoolsDeprecationWarning: setup.py install is deprecated. Use build and pip and other standards-based tools.

(This is pretty much what @astitcher said for Proton. Modern approach is to build tar.gz or wheel and then install it with pip.)

The `zip_safe` is not relevant when `--single-version-...` is in effect, looks like. So this can be removed from the PR.
Yet, I don't want to have no zip, so let's keep it consistent and say no to zip once for all.