Fixes #495: export default qd_buffer_t size for tests (#499)
This fixes the bug that required disabling the ASAN test for "one
definition rule" violations. By providing a public constant for the
default buffer size for use by the tests we can avoid attempting to
link the tests to the router libraries.

This patch also makes it harder to change the default buffer size by
inadvertantly calling a debug only interface.

Original fix by Andrew Stitcher.