ISSUE #754: work around possible Q2 egress stall
See ISSUE #754 (and #758). These failures appear to be cause by a
stall in the egress side. The test logs show that the egress HTTP/1
message is unable to completely receive a full stream body data
section (incomplete). Debug logging shows that the egress message has
Q2 in blocking state, which means the remaining octets of the stream
body data cannot be received until Q2 is unblocked. Q2 cannot be
unblocked since that can only occur when the entire stream body data
has arrived and sent out the raw connection.

This work-around disables Q2 on egress which should prevent the
stall. Unfortuantely I cannot reliably reproduce this stall locally so
time will tell if this workaround actually alleviates the stall.