ISSUE-773: refactor TCP and HTTP adaptor TLS configuration sequence
This patch refactors TLS into two separate classes: a qd_tls_domain_t
to represent the TLS domain and a qd_tls_t to managed the TLS session.

This patch also modifies the order of initialization of the TLS domain
and session have changed to better handle configuration and management
errors.