Issue #925 - Fix Proton Python main installation for Containerfile (#928)
The SYSINSTALL option for Proton stopped working on main. Python package needs to be used.

I either need to copy the tar.gz from the build container to the run container and install it there, or I need to install the tar.gz in the build container under --prefix.

This PR does the latter.