Fixes #1446: check for raw read close only after draining read buffers (#1447)
When polling the raw connection it is possible for
pn_raw_connection_is_read_closed() to return true when there are still
read buffers available. This can happen when the read buffers are not
completely drained due to flow control. The patch moves the check for
raw read closed to after the point where all read buffers have been
drained.

Closes #1446