Fixes #1456 - Implemented the administrative ACCESS_POINT and LINK
This also separates the closure-of-all-connections from the freeing of the server module so that connections can be closed at an earlier point of the shutdown sequence.

This also adds accepted-connections to the listener reference counts for a more orderly freeing of resources.

Original work by Ted Ross. Ported to amqp adaptor and obfuscated by kgiusti