Enable Clair security scanning on Quay
Since the router image was simplified, Quay reports scanning results,
but these are misleading. It turns out that while Clair (the scanning
tool used in Quay) can identify the packages shipped in the minimised
images, it needs extra information to identify vulnerabilities in
those packages; specifically, in UBI-based images, /root/buildinfo.
See
https://github.com/quay/claircore/blob/main/rhel/repositoryscanner.go
for reference.

This fixes that by copying the appropriate information from the
"packager" image. While we're at it, copy the repository information
too; this simplifies building further images on top of this one. Some
WORKDIR statements are superfluous so those are removed too.

Signed-off-by: Stephen Kitt <skitt@redhat.com>