Issue #1748: Implement inter-router connection group upgrade
This introduces the concept of a connection group ordinal. It
annotates inter-router connections. Its purpose is to assign a
precedence to connections within and inter-router connection
group. Those inter router connections with the highest ordinal become
the active routing path to a peer router. Connections within the group
that have a lower ordinal remain active but no new flows are sent over
them.

This is used to implement connection upgrade for certificate rotation.