Fixes #1753: Expire Listener-side connection on certificate rotation (#1763)
* Fixes #1753: Expire Listener-side connection on certificate rotation

When an AMQP listener's sslProfile oldestValidOrdinal attribute is
updated scan all connections that are children of that listener and
check if the connection's tlsOrdinal value is less than the new
oldestValidOrdinal value. If it is then close the connection.

Patch includes additional efforts to make the code fluffy:

1) Move qd_listener_t-specific code out of connection_manager.c and
move it to the qd_listener.c file.
2) Modify the qd_connection_t deferred callback API to allow an extra
context parameter to be passed to the callback. Hide the
implementation details by moving the data structure from a header file
into qd_connection.c.
3) Remove dead code.

* fixup: verify tlsOrdinal of new inter-router connections